requirements.txt
streamlit==1.28.1
google-generativeai==0.3.2
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
